pyproject.toml
[project]
name = "mediaflow-api"
version = "0.1.0"
description = "Media conversion API for images and video"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.21",
    "alembic>=1.13.0",
    "celery[redis]>=5.4.0",
    "redis>=5.1.0",
    "minio>=7.2.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.5.0",
    "python-magic>=0.4.27",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = false
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
